+++ README.md
@@ -7,26 +7,27 @@ This repo contains metadata related to LaunchDarkly SDKs.
 
 The data is intended for consumption by downstream products and services.
 
+It also contains a [Go module](./api) exposing the metadata for easier consumption by Go applications.
+
 | Data Product                             | Description                                                     | Format   |
 |------------------------------------------|-----------------------------------------------------------------|----------|
 | [Database](./metadata.sqlite3)           | Database containing data from which other products are derived. | sqlite3  |
 | [SDK Names](products/names.json)         | SDK friendly names for display.                                 | JSON     |
 | [SDK Releases](products/releases.json)   | SDK major/minor releases with EOL dates.                        | JSON     |
 | [SDK Types](products/types.json)         | SDK types for categorization.                                   | JSON     |
-| [SDK Features](products/features.json)   | SDK features, including version introduced/deprecated.          | JSON     |
 | [SDK Languages](products/languages.json) | Programming languages associated with SDKs.                     | JSON     |
 | [SDK Repos](products/repos.json)         | SDK source repositories                                         | JSON     |
 
 
 ## structure
 
-This repo contains an sqlite database containing a snapshot of SDK metadata
+This repo contains an [sqlite database](./metadata.sqlite3) containing a snapshot of SDK metadata
 fetched from individual repos.
 
-It also contains JSON files that are derived from the database. These are intended for
-consumption by downstream products and services.
+It also contains [JSON files](./products) that are derived from the database. These are intended for
+consumption by downstream products and services. 
 
-The JSON data products live in [`products`](./products) and the schemas for them live in [`schemas`](./schemas). 
+The JSON schemas for those products live in [`schemas`](./schemas). 
 
 ## tooling
 
@@ -37,9 +38,11 @@ Ensure that the JSON files are valid and formatted using `./scripts/ci/format-js
 To quickly create basic metadata for a new repo, run `./scripts/add-repo.sh <launchdarkly/repo-name>`. This will
 clone the repo, check-in an `.sdk_metadata.json` file, and create a PR.
 
+Some of the tooling lives in a [Go module](./tool) because it's too complex for a shell script.
+
 ## versioning policy
 
-The JSON products have an implicit 'v1' version.
+The JSON products have an implicit 'v1' version at this time.
 
 The sqlite database has no versioning policy and may change schemas at any time.
 
@@ -51,6 +54,12 @@ old ones so they remain up-to-date.
 3. Have new applications consume `/products/v2/whatever.json`
 4. Eventually, when/if all consumers have migrated, stop generating the old products. Don't delete them though, 
 just in case something is intermittently using them that we've forgotten about.
+
+The [API](./api) Go module should follow semver:
+- Patch for "bug fixes" to the JSON specs, or whenever metadata is updated in a backwards compatible way.
+- Minor for new data products (necessitating a new embedded table in the module)
+- Major for breaking changes to the API or data products. In the case of a major version, we need to append
+a new suffix to the module like `v2`.
 
 ### what necessitates a new major version
 - Changing the meaning of an existing field
