+++ api/go.mod
@@ -1,0 +1,11 @@
+module github.com/launchdarkly/sdk-meta/api
+
+go 1.20
+
+require github.com/stretchr/testify v1.9.0
+
+require (
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
+)
