+++ api/sdkmeta/data/languages.json
@@ -1,0 +1,108 @@
+{
+  "akamai-base": [
+    "JavaScript",
+    "TypeScript"
+  ],
+  "akamai-edgekv": [
+    "JavaScript",
+    "TypeScript"
+  ],
+  "android": [
+    "Java",
+    "Kotlin"
+  ],
+  "apex-server-sdk": [
+    "Apex"
+  ],
+  "cloudflare": [
+    "JavaScript",
+    "TypeScript"
+  ],
+  "cpp-client-sdk": [
+    "C++",
+    "C"
+  ],
+  "cpp-server-sdk": [
+    "C++",
+    "C"
+  ],
+  "dotnet-client-sdk": [
+    "C#"
+  ],
+  "dotnet-server-sdk": [
+    "C#"
+  ],
+  "electron": [
+    "JavaScript",
+    "TypeScript"
+  ],
+  "erlang-server-sdk": [
+    "Erlang",
+    "Elixir"
+  ],
+  "flutter-client-sdk": [
+    "Dart"
+  ],
+  "go-server-sdk": [
+    "Go"
+  ],
+  "haskell-server-sdk": [
+    "Haskell"
+  ],
+  "java-server-sdk": [
+    "Java"
+  ],
+  "js-client-sdk": [
+    "JavaScript",
+    "TypeScript"
+  ],
+  "lua-server-sdk": [
+    "Lua"
+  ],
+  "node-client-sdk": [
+    "JavaScript",
+    "TypeScript"
+  ],
+  "node-server": [
+    "JavaScript",
+    "TypeScript"
+  ],
+  "php-server-sdk": [
+    "PHP"
+  ],
+  "python-server-sdk": [
+    "Python"
+  ],
+  "react-client-sdk": [
+    "JavaScript",
+    "TypeScript"
+  ],
+  "react-native": [
+    "JavaScript",
+    "TypeScript"
+  ],
+  "relay-proxy": [
+    "Go"
+  ],
+  "roku": [
+    "BrightScript"
+  ],
+  "ruby-server-sdk": [
+    "Ruby"
+  ],
+  "rust-server-sdk": [
+    "Rust"
+  ],
+  "swift-client-sdk": [
+    "Swift",
+    "Objective-C"
+  ],
+  "vercel": [
+    "JavaScript",
+    "TypeScript"
+  ],
+  "vue": [
+    "JavaScript",
+    "TypeScript"
+  ]
+}
