+++ api/sdkmeta/data/names.json
@@ -1,0 +1,32 @@
+{
+  "akamai-base": "Akamai SDK",
+  "akamai-edgekv": "Akamai SDK for EdgeKV",
+  "android": "Android SDK",
+  "apex-server-sdk": "Apex Server SDK",
+  "cloudflare": "Cloudflare SDK",
+  "cpp-client-sdk": "C++ Client SDK",
+  "cpp-server-sdk": "C++ Server SDK",
+  "dotnet-client-sdk": ".NET Client SDK",
+  "dotnet-server-sdk": ".NET Server SDK",
+  "electron": "Electron SDK",
+  "erlang-server-sdk": "Erlang Server SDK",
+  "flutter-client-sdk": "Flutter Client SDK",
+  "go-server-sdk": "Go Server SDK",
+  "haskell-server-sdk": "Haskell Server SDK",
+  "java-server-sdk": "Java Server SDK",
+  "js-client-sdk": "JavaScript Client SDK",
+  "lua-server-sdk": "Lua Server SDK",
+  "node-client-sdk": "Node.js Client SDK",
+  "node-server": "Node.js Server SDK",
+  "php-server-sdk": "PHP SDK",
+  "python-server-sdk": "Python Server SDK",
+  "react-client-sdk": "React Web SDK",
+  "react-native": "React Native SDK",
+  "relay-proxy": "Relay Proxy",
+  "roku": "Roku SDK",
+  "ruby-server-sdk": "Ruby Server SDK",
+  "rust-server-sdk": "Rust Server SDK",
+  "swift-client-sdk": "iOS SDK",
+  "vercel": "Vercel Edge SDK",
+  "vue": "Vue SDK"
+}
