+++ api/sdkmeta/data/repos.json
@@ -1,0 +1,92 @@
+{
+  "akamai-base": {
+    "github": "launchdarkly/js-core"
+  },
+  "akamai-edgekv": {
+    "github": "launchdarkly/js-core"
+  },
+  "android": {
+    "github": "launchdarkly/android-client-sdk"
+  },
+  "apex-server-sdk": {
+    "github": "launchdarkly/apex-server-sdk"
+  },
+  "cloudflare": {
+    "github": "launchdarkly/js-core"
+  },
+  "cpp-client-sdk": {
+    "github": "launchdarkly/cpp-sdks"
+  },
+  "cpp-server-sdk": {
+    "github": "launchdarkly/cpp-sdks"
+  },
+  "dotnet-client-sdk": {
+    "github": "launchdarkly/dotnet-core"
+  },
+  "dotnet-server-sdk": {
+    "github": "launchdarkly/dotnet-core"
+  },
+  "electron": {
+    "github": "launchdarkly/electron-client-sdk"
+  },
+  "erlang-server-sdk": {
+    "github": "launchdarkly/erlang-server-sdk"
+  },
+  "flutter-client-sdk": {
+    "github": "launchdarkly/flutter-client-sdk"
+  },
+  "go-server-sdk": {
+    "github": "launchdarkly/go-server-sdk"
+  },
+  "haskell-server-sdk": {
+    "github": "launchdarkly/haskell-server-sdk"
+  },
+  "java-server-sdk": {
+    "github": "launchdarkly/java-core"
+  },
+  "js-client-sdk": {
+    "github": "launchdarkly/js-client-sdk"
+  },
+  "lua-server-sdk": {
+    "github": "launchdarkly/lua-server-sdk"
+  },
+  "node-client-sdk": {
+    "github": "launchdarkly/node-client-sdk"
+  },
+  "node-server": {
+    "github": "launchdarkly/js-core"
+  },
+  "php-server-sdk": {
+    "github": "launchdarkly/php-server-sdk"
+  },
+  "python-server-sdk": {
+    "github": "launchdarkly/python-server-sdk"
+  },
+  "react-client-sdk": {
+    "github": "launchdarkly/react-client-sdk"
+  },
+  "react-native": {
+    "github": "launchdarkly/js-core"
+  },
+  "relay-proxy": {
+    "github": "launchdarkly/ld-relay"
+  },
+  "roku": {
+    "github": "launchdarkly/roku-client-sdk"
+  },
+  "ruby-server-sdk": {
+    "github": "launchdarkly/ruby-server-sdk"
+  },
+  "rust-server-sdk": {
+    "github": "launchdarkly/rust-server-sdk"
+  },
+  "swift-client-sdk": {
+    "github": "launchdarkly/ios-client-sdk"
+  },
+  "vercel": {
+    "github": "launchdarkly/js-core"
+  },
+  "vue": {
+    "github": "launchdarkly/vue-client-sdk"
+  }
+}
