+++ api/sdkmeta/data/types.json
@@ -1,0 +1,32 @@
+{
+  "akamai-base": "edge",
+  "akamai-edgekv": "edge",
+  "android": "client-side",
+  "apex-server-sdk": "server-side",
+  "cloudflare": "edge",
+  "cpp-client-sdk": "client-side",
+  "cpp-server-sdk": "server-side",
+  "dotnet-client-sdk": "client-side",
+  "dotnet-server-sdk": "server-side",
+  "electron": "client-side",
+  "erlang-server-sdk": "server-side",
+  "flutter-client-sdk": "client-side",
+  "go-server-sdk": "server-side",
+  "haskell-server-sdk": "server-side",
+  "java-server-sdk": "server-side",
+  "js-client-sdk": "client-side",
+  "lua-server-sdk": "server-side",
+  "node-client-sdk": "client-side",
+  "node-server": "server-side",
+  "php-server-sdk": "server-side",
+  "python-server-sdk": "server-side",
+  "react-client-sdk": "client-side",
+  "react-native": "client-side",
+  "relay-proxy": "relay",
+  "roku": "client-side",
+  "ruby-server-sdk": "server-side",
+  "rust-server-sdk": "server-side",
+  "swift-client-sdk": "client-side",
+  "vercel": "edge",
+  "vue": "client-side"
+}
