+++ api/sdkmeta/sdkmeta.go
@@ -1,4 +1,4 @@
-package products
+package sdkmeta
 
 import (
 	_ "embed"
@@ -6,13 +6,13 @@ import (
 	"time"
 )
 
-//go:embed names.json
+//go:embed data/names.json
 var namesJSON []byte
 
 // Names is a map of SDK IDs to display names.
 var Names map[string]string
 
-//go:embed repos.json
+//go:embed data/repos.json
 var reposJSON []byte
 
 // Repo contains the location of an SDK.
@@ -24,13 +24,13 @@ type Repo struct {
 // Repos is a map of SDK IDs to repository information.
 var Repos map[string]Repo
 
-//go:embed languages.json
+//go:embed data/languages.json
 var languagesJSON []byte
 
 // Languages is a map of SDK IDs to supported languages.
 var Languages map[string][]string
 
-//go:embed types.json
+//go:embed data/types.json
 var typesJSON []byte
 
 // Type represents the most common use-case for an SDK.
@@ -50,7 +50,7 @@ const (
 // Types is a map of SDK IDs to SDK types.
 var Types map[string]Type
 
-//go:embed releases.json
+//go:embed data/releases.json
 var releasesJSON []byte
 
 type Release struct {
